+++ github3/github.py
@@ -779,11 +779,23 @@ class GitHub(GitHubCore):
         url = self._build_url('user', 'starred')
         return self._iter(int(number), url, Repository, params, etag)
 
-    def subscriptions(self, login=None, number=-1, etag=None):
-        """Iterate over repositories subscribed to by ``login`` or the
-        authenticated user.
+    @requires_auth
+    def subscriptions(self, number=-1, etag=None):
+        """Iterate over repositories subscribed to by the authenticated user.
 
-        :param str login: (optional), name of user whose subscriptions you want
+        :param int number: (optional), number of repositories to return.
+            Default: -1 returns all repositories
+        :param str etag: (optional), ETag from a previous request to the same
+            endpoint
+        :returns: generator of :class:`Repository <github3.repos.Repository>`
+        """
+        url = self._build_url('user', 'subscriptions')
+        return self._iter(int(number), url, Repository, etag=etag)
+
+    def subscriptions_for(self, login, number=-1, etag=None):
+        """Iterate over repositories subscribed to by ``login``.
+
+        :param str login: , name of user whose subscriptions you want
             to see
         :param int number: (optional), number of repositories to return.
             Default: -1 returns all repositories
@@ -791,10 +803,7 @@ class GitHub(GitHubCore):
             endpoint
         :returns: generator of :class:`Repository <github3.repos.Repository>`
         """
-        if login:
-            return self.user(login).iter_subscriptions()
-
-        url = self._build_url('user', 'subscriptions')
+        url = self._build_url('users', str(login), 'subscriptions')
         return self._iter(int(number), url, Repository, etag=etag)
 
     def user_repos(self, login, type=None, sort=None, direction=None,